refactoring

+++ src/myimg/myimg.h
@@ -4,6 +4,6 @@ int set_rgba (int, int, int, int, int, int);
 int paint_img_background (int, int, int, int);
 int paint_img_border (int, int, int, int);
 int paint_x_axis (float, float, float, float, int, int, int, int);
-int paint_y_axis (float, float, float, int, float, int, int, int, int);
+int paint_y_axis (float, float, float, float, int, int, int, int, int);
 int write_img (char *);
 void free_img ();
